# openapi.yaml
openapi: 3.0.3
info:
  title: Cerberus Network Sentinel API
  version: 1.0.0
  description: >
    # Welcome to the Cerberus Network Sentinel API


    Cerberus Sentinel is an automated network security auditing and asset
    inventory daemon built in Go. It continuously monitors network subnets,
    profiles active devices and transport services, correlates banner metadata with
    known Common Vulnerabilities and Exposures (CVEs), and exposes real-time
    telemetry through a RESTful JSON interface.


    ### How this documentation works


    This API reference is automatically generated from the OpenAPI specification
    uploaded to GitBook. GitBook parses the schema to render interactive endpoint
    explorers, schema definitions, and executable code samples.


    Key architectural features of this API:


    - **Page groups** — Tags are organized into hierarchical sections (*Inventory*,
    *Vulnerability Analysis*, *Scan Controls*, and *System Telemetry*), keeping related
    endpoints naturally structured.

    - **Auto-split sections** — `##` headings inside `info.description` are
    automatically partitioned into distinct subpages in the sidebar navigation.

    - **Reactive & Scheduled Operations** — The sentinel coordinates background
    goroutines (30-minute discovery sweeps, 10-minute vulnerability audits) as well
    as immediate on-demand scans.

    - **Interactive Code Samples** — Every endpoint features hand-crafted code
...

...


...


    ### Servers


...

...

...

...

...


...

...

...

...

...

...
servers:
...
paths:
  /api/status:
    get:
      summary: Retrieve system telemetry and status
      description: >
        Returns real-time runtime metrics for the Cerberus daemon, including
        process uptime, aggregated asset counts, current scheduler running states,
        and timestamps for the next planned discovery and vulnerability sweeps.
      tags:
        - system
      operationId: getSystemStatus
      responses:
        '200':
          description: Live sentinel health and scheduler metrics.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatus'
              example:
                uptime: 1h42m15s
                uptime_seconds: 6135
                active_devices: 5
                total_devices: 5
                total_services: 8
                total_vulnerabilities: 102
                discovery_next_run: '2026-09-25T12:00:00Z'
                vuln_next_run: '2026-09-25T11:50:00Z'
                is_discovery_running: false
                is_vuln_running: false
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-codeSamples:
        - lang: cURL
          label: CLI
          source: |
            curl -L http://localhost:8080/api/status
        - lang: JavaScript
          label: Fetch
          source: |
            const res = await fetch('http://localhost:8080/api/status');
            const status = await res.json();
            console.log(status);
        - lang: Python
          label: Python
          source: |
            import requests
            r = requests.get('http://localhost:8080/api/status')
            print(r.json())
...
  /api/devices/{id}:
    get:
      summary: Retrieve detailed device profile
      description: >
        Fetches complete forensic details for a single host by its numeric identifier,
        including all identified transport-layer ports, application banners, and
        correlated CVE security vulnerabilities.
      tags:
        - devices
      operationId: getDeviceById
      parameters:
        - name: id
          in: path
          required: true
          description: Unique integer ID of the device.
          schema:
            type: integer
            format: int64
          example: 2
      responses:
        '200':
          description: Full device entity with nested services and vulnerabilities.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
              example:
                id: 2
                ip: 172.28.0.30
                mac: 02:42:AC:1C:00:1E
                hostname: target-dvwa
                vendor: Docker Network
                status: up
                first_seen: '2026-09-25T10:00:00Z'
                last_seen: '2026-09-25T11:30:00Z'
                services_count: 1
                vulnerabilities_count: 1
                services:
                  - id: 1
                    device_id: 2
                    device_ip: 172.28.0.30
                    port: 80
                    protocol: tcp
                    service_name: http
                    product: Apache httpd
                    version: 2.4.25
                    state: open
                    updated_at: '2026-09-25T11:30:00Z'
                vulnerabilities:
                  - id: 1
                    device_id: 2
                    device_ip: 172.28.0.30
                    service_port: 80
                    cve_id: CVE-2026-44631
                    title: Vulnerability CVE-2026-44631 on port 80
                    severity: CRITICAL
                    cvss_score: 9.8
                    description: Identified via vulners script
                    reference_url: https://vulners.com/cve/CVE-2026-44631
                    detected_at: '2026-09-25T11:30:00Z'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-codeSamples:
        - lang: cURL
          label: CLI
          source: |
            curl -L http://localhost:8080/api/devices/2
        - lang: JavaScript
          label: Fetch
          source: |
            const res = await fetch('http://localhost:8080/api/devices/2');
            const device = await res.json();
            console.log(device);
        - lang: Python
          label: Python
          source: |
            import requests
            r = requests.get('http://localhost:8080/api/devices/2')
            print(r.json())
...
  /api/timeline:
    get:
      summary: Retrieve device activity periods and scan line ticks
      description: >
        Returns time-series scan execution ticks alongside segmented continuous activity
        periods for all tracked devices. Activity periods automatically segment across
        liveness transitions (up <-> down) and vulnerability posture changes (severity shifts),
        enabling direct rendering of interactive Gantt-style timeline graphs and scan lines.
      tags:
        - timeline
      operationId: getTimeline
      parameters:
        - name: from
          in: query
          required: false
          description: RFC3339 timestamp or duration (e.g. 24h, 7d) defining the start of the time window. Defaults to 24 hours ago.
          schema:
            type: string
            example: "2026-09-30T06:00:00Z"
        - name: to
          in: query
          required: false
          description: RFC3339 timestamp defining the end of the time window. Defaults to current timestamp.
          schema:
            type: string
            example: "2026-09-30T12:00:00Z"
        - name: status
          in: query
          required: false
          description: Filter devices by their current status.
          schema:
            type: string
            enum: [up, down]
        - name: vulnerable_only
          in: query
          required: false
          description: If true, filters response to include only devices with detected vulnerabilities during the time window.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Timeline dataset containing scan ticks and device activity periods.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-codeSamples:
        - lang: cURL
          label: CLI
          source: |
            curl -L "http://localhost:8080/api/timeline?from=24h&vulnerable_only=false"
        - lang: JavaScript
          label: Fetch
          source: |
            const res = await fetch('http://localhost:8080/api/timeline?from=24h');
            const data = await res.json();
            console.log(data.devices, data.scan_ticks);
        - lang: Python
          label: Python
          source: |
            import requests
            r = requests.get('http://localhost:8080/api/timeline', params={'from': '24h'})
            print(r.json())
...
  /api/scans/vulnerability/trigger:
    post:
      summary: Trigger on-demand vulnerability audit
      description: >
        Asynchronously initiates an immediate vulnerability assessment sweep across
        all active endpoints currently registered in the database. Returns immediately
        with HTTP 202 Accepted.
      tags:
        - scans
      operationId: triggerVulnerabilityScan
      responses:
        '202':
          description: Vulnerability audit task dispatched to background worker.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanTriggerResponse'
              example:
                message: Vulnerability audit initiated in background
                status: running
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-codeSamples:
        - lang: cURL
          label: CLI
          source: |
            curl -L -X POST http://localhost:8080/api/scans/vulnerability/trigger
        - lang: JavaScript
          label: Fetch
          source: |
            const res = await fetch('http://localhost:8080/api/scans/vulnerability/trigger', {
              method: 'POST'
            });
            const data = await res.json();
            console.log(data);
        - lang: Python
          label: Python
          source: |
            import requests
            r = requests.post('http://localhost:8080/api/scans/vulnerability/trigger')
            print(r.json())
  /api/export/json:
    get:
      summary: Export complete network inventory snapshot
      description: >
        Emits a consolidated, structured JSON export document containing all tracked
        network devices, their active services, and all correlated CVE vulnerability findings.
        Supplied with a downloadable attachment header (`Content-Disposition`).
      tags:
        - system
      operationId: exportAllData
      responses:
        '200':
          description: Complete serialized inventory snapshot.
          headers:
            Content-Disposition:
              schema:
                type: string
              example: attachment; filename="cerberus-export-20260925-115500.json"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportData'
        '500':
          $ref: '#/components/responses/InternalServerError'
      x-hideTryItPanel: false
      x-codeSamples:
        - lang: cURL
          label: CLI
          source: |
            curl -L -O -J http://localhost:8080/api/export/json
        - lang: JavaScript
          label: Fetch
          source: |
            const res = await fetch('http://localhost:8080/api/export/json');
            const backup = await res.json();
            console.log(`Total devices: ${backup.total_devices}`);
        - lang: Python
          label: Python
          source: |
            import requests
            r = requests.get('http://localhost:8080/api/export/json')
            with open('cerberus_export.json', 'w') as f:
                f.write(r.text)
...
components:
  schemas:
    Device:
      type: object
      required:
        - id
        - ip
        - status
        - first_seen
        - last_seen
      properties:
        id:
          type: integer
          format: int64
          description: Unique surrogate identifier for the device.
          example: 1
        ip:
          type: string
          description: IPv4 or IPv6 address.
          example: 172.28.0.10
        mac:
          type: string
          description: Physical MAC hardware address.
          example: 02:42:AC:1C:00:0A
        hostname:
          type: string
          description: Resolved PTR or NetBIOS hostname.
          example: target-nginx
        vendor:
          type: string
          description: Hardware manufacturer inferred from MAC OUI prefix.
          example: Docker Network
        status:
          type: string
          enum: [up, down]
          description: Operational liveness state.
          example: up
        first_seen:
          type: string
          format: date-time
          description: Initial discovery timestamp.
        last_seen:
          type: string
          format: date-time
          description: Most recent active response timestamp.
        services_count:
          type: integer
          description: Number of open services detected on the host.
          example: 1
        vulnerabilities_count:
          type: integer
          description: Number of vulnerabilities correlated with the host.
          example: 0
        services:
          type: array
          description: Detailed list of active services (included on detail queries).
          items:
            $ref: '#/components/schemas/Service'
        vulnerabilities:
          type: array
          description: Detailed list of correlated vulnerabilities (included on detail queries).
          items:
            $ref: '#/components/schemas/Vulnerability'
...
    Service:
      type: object
      required:
        - id
        - device_id
        - port
        - protocol
        - state
        - updated_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique service record identifier.
          example: 1
        device_id:
          type: integer
          format: int64
          description: Parent device foreign key identifier.
          example: 2
        device_ip:
          type: string
          description: IP address of the parent device.
          example: 172.28.0.30
        port:
          type: integer
          description: Transport-layer port number.
          example: 80
        protocol:
          type: string
          enum: [tcp, udp]
          description: Layer-4 transport protocol.
          example: tcp
        service_name:
          type: string
          description: Canonical service name.
          example: http
        product:
          type: string
          description: Interrogated software application name.
          example: Apache httpd
        version:
          type: string
          description: Interrogated software version string.
          example: 2.4.25
        state:
          type: string
          enum: [open, filtered, closed]
          description: Network socket state.
          example: open
        updated_at:
          type: string
          format: date-time
          description: Timestamp of most recent service interrogation.
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/Vulnerability'
...
    VulnerabilitySeverity:
      type: string
      enum:
        - CRITICAL
        - HIGH
        - MEDIUM
        - LOW
        - UNKNOWN
      x-enumDescriptions:
        CRITICAL: CVSS score >= 9.0. Exploitation allows remote code execution or complete system takeover.
        HIGH: CVSS score between 7.0 and 8.9. Severe vulnerabilities requiring urgent mitigation.
        MEDIUM: CVSS score between 4.0 and 6.9. Moderately complex or constrained vulnerabilities.
        LOW: CVSS score between 0.1 and 3.9. Minor information disclosures or low-impact weaknesses.
        UNKNOWN: Informational or unrated security advisory.
    Vulnerability:
      type: object
      required:
        - id
        - device_id
        - service_port
        - cve_id
        - severity
        - cvss_score
        - detected_at
      properties:
        id:
          type: integer
          format: int64
          description: Unique vulnerability record identifier.
          example: 1
        device_id:
          type: integer
          format: int64
          description: Affected device foreign key identifier.
          example: 2
        device_ip:
          type: string
          description: IP address of the affected device.
          example: 172.28.0.30
        service_port:
          type: integer
          description: Open port number through which the vulnerability is exposed.
          example: 80
        cve_id:
          type: string
          description: Standard CVE vulnerability identifier.
          example: CVE-2026-44631
        title:
          type: string
          description: Summary headline of the vulnerability.
          example: Vulnerability CVE-2026-44631 on port 80
        severity:
          $ref: '#/components/schemas/VulnerabilitySeverity'
        cvss_score:
          type: number
          format: double
          description: CVSS v3.x numerical base score (0.0 to 10.0).
          example: 9.8
        description:
          type: string
          description: Technical summary of the vulnerability.
          example: Identified via vulners script
        reference_url:
          type: string
          format: uri
          description: Link to authoritative NIST NVD or Vulners security advisory.
          example: https://vulners.com/cve/CVE-2026-44631
        detected_at:
          type: string
          format: date-time
          description: Initial detection timestamp.
...
    SystemStatus:
      type: object
      required:
        - uptime
        - uptime_seconds
        - active_devices
        - total_devices
        - total_services
        - total_vulnerabilities
        - is_discovery_running
        - is_vuln_running
      properties:
        uptime:
          type: string
          description: Human-readable daemon uptime duration.
          example: 1h42m15s
        uptime_seconds:
          type: integer
          format: int64
          description: Total uptime in seconds.
          example: 6135
        active_devices:
          type: integer
          description: Total count of devices currently marked as up.
          example: 5
        total_devices:
          type: integer
          description: Total count of historical devices in database.
          example: 5
        total_services:
          type: integer
          description: Total count of active services identified.
          example: 8
        total_vulnerabilities:
          type: integer
          description: Total count of active vulnerabilities across all hosts.
          example: 102
        discovery_next_run:
          type: string
          format: date-time
          description: Next planned discovery sweep timestamp.
        vuln_next_run:
          type: string
          format: date-time
          description: Next planned bulk vulnerability audit timestamp.
        is_discovery_running:
          type: boolean
          description: True if a subnet discovery sweep is currently executing.
          example: false
        is_vuln_running:
          type: boolean
          description: True if a vulnerability audit sweep is currently executing.
          example: false
    ScanTriggerResponse:
      type: object
      required:
        - message
        - status
      properties:
        message:
          type: string
          description: Informational status message.
          example: Discovery scan initiated in background
        status:
          type: string
          description: Execution state.
          example: running
    ExportData:
      type: object
      required:
        - exported_at
        - network_cidr
        - total_devices
        - total_vulnerabilities
        - devices
      properties:
        exported_at:
          type: string
          format: date-time
          description: Timestamp when snapshot was exported.
        network_cidr:
          type: string
          description: Configured network CIDR.
          example: 172.28.0.0/24
        total_devices:
          type: integer
          description: Number of exported devices.
          example: 5
        total_vulnerabilities:
          type: integer
          description: Number of exported vulnerabilities.
          example: 102
        devices:
          type: array
          items:
            $ref: '#/components/schemas/Device'
    TimelineResponse:
      type: object
      required:
        - time_window
        - scan_ticks
        - devices
      properties:
        time_window:
          $ref: '#/components/schemas/TimelineTimeWindow'
        scan_ticks:
          type: array
          items:
            $ref: '#/components/schemas/ScanTick'
        devices:
          type: array
          items:
            $ref: '#/components/schemas/TimelineDevice'
    TimelineTimeWindow:
      type: object
      required:
        - from
        - to
        - total_scans
        - total_devices
      properties:
        from:
          type: string
          format: date-time
          example: "2026-09-30T06:00:00Z"
        to:
          type: string
          format: date-time
          example: "2026-09-30T12:00:00Z"
        total_scans:
          type: integer
          example: 7
        total_devices:
          type: integer
          example: 4
    ScanTick:
      type: object
      required:
        - scan_id
        - timestamp
        - scan_type
        - status
      properties:
        scan_id:
          type: integer
          example: 101
        timestamp:
          type: string
          format: date-time
          example: "2026-09-30T06:00:00Z"
        scan_type:
          type: string
          example: discovery
        status:
          type: string
          example: completed
        target_spec:
          type: string
          example: "172.28.0.0/24"
        hosts_found:
          type: integer
          example: 3
        duration_ms:
          type: integer
          example: 1420
    TimelineDevice:
      type: object
      required:
        - device_id
        - ip
        - current_status
        - periods
      properties:
        device_id:
          type: integer
          example: 2
        ip:
          type: string
          example: "172.28.0.20"
        mac:
          type: string
          example: "02:42:AC:1C:00:14"
        hostname:
          type: string
          example: target-nginx
        vendor:
          type: string
          example: Docker Inc.
        current_status:
          type: string
          example: up
        periods:
          type: array
          items:
            $ref: '#/components/schemas/TimelinePeriod'
    TimelinePeriod:
      type: object
      required:
        - started_at
        - ended_at
        - status
        - is_vulnerable
        - vulns_count
        - max_severity
        - is_ongoing
      properties:
        started_at:
          type: string
          format: date-time
          example: "2026-09-30T08:10:00Z"
        ended_at:
          type: string
          format: date-time
          example: "2026-09-30T12:00:00Z"
        status:
          type: string
          example: up
        is_vulnerable:
          type: boolean
          example: true
        vulns_count:
          type: integer
          example: 2
        max_severity:
          type: string
          example: HIGH
        is_ongoing:
          type: boolean
          example: true
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Description of the error that occurred.
          example: device 42 not found
  responses:
    BadRequest:
      description: The request contains invalid parameters or malformed identifiers.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: invalid device id
    NotFound:
      description: The requested resource could not be found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: device 99 not found
    InternalServerError:
      description: The server encountered an unexpected error processing the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: database operation failed